done ch04

+++ bin/rsync_container_to_host.sh
@@ -8,31 +8,53 @@ SRC="${CONTAINER_NAME}:/var/www/html/wp-content"
 DEST=$(pwd)/wp/wp-content
 EXCLUDE="$(pwd)/etc/exclude-file.txt"
 
-while getopts t: OPT
+usage() {
+	echo "Usage: ${CMDNAME} [-t, --target all|theme|uploads|plugins] [-d, --delete true|false] delete files in container" 1>&2
+	exit 1
+}
+
+while getopts t:d:h OPT
 do
 	case ${OPT} in
-		"t" ) FLG_T="TRUE" ; VALUE_T="${OPTARG}";;
-			* ) echo "Usage: ${CMDNAME} [-t all|theme|uploads|plugins]" 1>&2
-			exit 1;;
+		"t" | "target" )
+			FLG_T="TRUE"
+			VALUE_T="${OPTARG}"
+		;;
+		"d" | "delete" )
+			FLG_D="TRUE"
+			VALUE_D="${OPTARG}"
+		;;
+		"h" ) usage;;
+		\? ) usage;;
 	esac
 done
 
-if [ "${FLG_T}" = "TRUE" ] && [ "${VALUE_T}" = "all" ]; then
-	rsync --blocking-io -e 'docker exec -i' -ahv --delete --exclude-from=${EXCLUDE} ${SRC}/themes/res ${DEST}/themes/
+if [ "${FLG_T}" = "TRUE" ] && [ "${VALUE_T}" = "all" ] && [ "${FLG_D}" = "TRUE" ] && [ "${VALUE_D}" = "TRUE" ]; then
+	rsync --blocking-io -e 'docker exec -i' -ahv --delete --exclude-from=${EXCLUDE} ${SRC}/themes ${DEST}/
 	rsync --blocking-io -e 'docker exec -i' -ahv --delete --exclude-from=${EXCLUDE} ${SRC}/uploads ${DEST}/
 	rsync --blocking-io -e 'docker exec -i' -ahv --delete --exclude-from=${EXCLUDE} ${SRC}/plugins ${DEST}/
+else
+	rsync --blocking-io -e 'docker exec -i' -ahv --exclude-from=${EXCLUDE} ${SRC}/themes ${DEST}/
+	rsync --blocking-io -e 'docker exec -i' -ahv --exclude-from=${EXCLUDE} ${SRC}/uploads ${DEST}/
+	rsync --blocking-io -e 'docker exec -i' -ahv --exclude-from=${EXCLUDE} ${SRC}/plugins ${DEST}/
 fi
 
 if [ "${FLG_T}" = "TRUE" ] && [ "${VALUE_T}" = "theme" ]; then
-	rsync --blocking-io -e 'docker exec -i' -ahv --delete --exclude-from=${EXCLUDE} ${SRC}/themes/res ${DEST}/themes/
+	rsync --blocking-io -e 'docker exec -i' -ahv --delete --exclude-from=${EXCLUDE} ${SRC}/themes ${DEST}/themes
+else
+	rsync --blocking-io -e 'docker exec -i' -ahv --exclude-from=${EXCLUDE} ${SRC}/themes ${DEST}/
 fi
 
 if [ "${FLG_T}" = "TRUE" ] && [ "${VALUE_T}" = "uploads" ]; then
 	rsync --blocking-io -e 'docker exec -i' -ahv --delete --exclude-from=${EXCLUDE} ${SRC}/uploads ${DEST}/
+else
+	rsync --blocking-io -e 'docker exec -i' -ahv --exclude-from=${EXCLUDE} ${SRC}/uploads ${DEST}/
 fi
 
 if [ "${FLG_T}" = "TRUE" ] && [ "${VALUE_T}" = "plugins" ]; then
 	rsync --blocking-io -e 'docker exec -i' -ahv --delete --exclude-from=${EXCLUDE} ${SRC}/plugins ${DEST}/
+else
+	rsync --blocking-io -e 'docker exec -i' -ahv --exclude-from=${EXCLUDE} ${SRC}/plugins ${DEST}/
 fi
